cat embedding vars

+++ layers/word_embedding.py
@@ -12,8 +12,7 @@ class WordEmbedding(nn.Module):
             self.embedding.weight = nn.Parameter(args.pre_embd_w, requires_grad=is_train_embd)
         
     def forward(self, x):
-        x = self.embedding(x)
+        x = self.embedding(x).type(torch.FloatTensor) # TODO
         out = F.relu(x)
-        # print('out', out.size())
 
         return out
